pyproject.toml
[tool.poetry]
name = "senior-common-room"
version = "0.1.0"
description = ""
authors = ["Mark Hall <mark.hall@open.ac.uk>"]

[tool.poetry.dependencies]
python = "^3.9"
tornado = "^6.1"
click = "^8.0.1"
SQLAlchemy = "^1.4.18"
sqlalchemy-json = "^0.4.0"
PyYAML = "^5.4.1"
asyncpg = {version = "^0.23.0", optional = true}
Cerberus = "^1.3.4"
Pillow = "^8.3.1"
asyncio-mqtt = "^0.9.1"
PyJWT = "^2.1.0"

[tool.poetry.dev-dependencies]

[tool.poetry.extras]
postgresql = ["asyncpg"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
